# hotel-backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  hotel-backend_internal_models.Role:
    enum:
    - admin
    - receptionist
    - housekeeping
    type: string
    x-enum-varnames:
    - RoleAdmin
    - RoleReceptionist
    - RoleHousekeeping
...
  hotel-backend_internal_models.UserRegisterRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 6
        type: string
      role:
        $ref: '#/definitions/hotel-backend_internal_models.Role'
    required:
    - email
    - name
    - password
    - role
    type: object
  hotel-backend_internal_models.UserUpdateRequest:
    properties:
      email:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      role:
        $ref: '#/definitions/hotel-backend_internal_models.Role'
    type: object
host: localhost:8080
info:
  contact:
    email: support@hotel-system.com
    name: API Support
  description: This is the API documentation for the Hotel Management System.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Hotel Management System API
  version: "1.0"
paths:
...
  /staff:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all staff members (Admin only)
      produces:
      - application/json
      responses:
        "200":
          description: Staff users retrieved
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - Admin only
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all staff users
      tags:
      - Staff
    post:
      consumes:
      - application/json
      description: Register a new staff user (Admin only)
      parameters:
      - description: Staff user details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/hotel-backend_internal_models.UserRegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Staff member created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - Admin only
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new staff user
      tags:
      - Staff
  /staff/{id}:
    delete:
      consumes:
      - application/json
      description: Remove a staff user by ID (Admin only)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Staff user deleted successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid user ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - Admin only
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to delete staff user
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete staff user
      tags:
      - Staff
    get:
      consumes:
      - application/json
      description: Retrieve details of a specific staff user by ID (Admin only)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Staff user retrieved
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid user ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - Admin only
          schema:
            additionalProperties: true
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get staff user by ID
      tags:
      - Staff
    put:
      consumes:
      - application/json
      description: Update staff user details by ID (Admin only)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User update data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/hotel-backend_internal_models.UserUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Staff user updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request or validation error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - Admin only
          schema:
            additionalProperties: true
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update staff user
      tags:
      - Staff
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: 'Enter your JWT token as: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"